@@ -325,7 +325,7 @@ class AdminBaseController
             }
 
             $isMime = strstr($type, '/');
-            $find   = str_replace(['.', '*'], ['\.', '.*'], $type);
+            $find   = str_replace(['.', '*', '+'], ['\.', '.*', '\+'], $type);
 
             if ($isMime) {
                 $match = preg_match('#' . $find . '$#', $mime);
@@ -737,8 +737,8 @@ class AdminBaseController
         // Process previously uploaded files for the current URI
         // and finally store them. Everything else will get discarded
         $queue = $this->admin->session()->getFlashObject('files-upload');
-        $queue = $queue[base64_encode($this->grav['uri']->url())];
         if (is_array($queue)) {
+            $queue = $queue[base64_encode($this->grav['uri']->url())];
             foreach ($queue as $key => $files) {
                 foreach ($files as $destination => $file) {
                     if (!rename($file['tmp_name'], $destination)) {
@@ -943,11 +943,11 @@ class AdminBaseController
             $settings = (object)$blueprints->schema()->getProperty($field);
         } else {
             $page = null;
-            if ($type === 'user') {
-                $settings = (object)$this->admin->blueprints($blueprint)->schema()->getProperty($field);
+            if ($type === 'themes') {
+                $obj = $this->grav[$type]->get(Utils::substrToString($blueprint, '/')); //here
+                $settings = (object) $obj->blueprints()->schema()->getProperty($field);
             } else {
-                $obj = $this->grav[$type]->get(Utils::substrToString($blueprint, '/'));
-                $settings = (object)$obj->blueprints()->schema()->getProperty($field);
+                $settings = (object)$this->admin->blueprints($blueprint)->schema()->getProperty($field);
             }
         }
 
