@@ -24,6 +24,7 @@ use Grav\Common\User\Interfaces\UserInterface;
 use Grav\Common\Utils;
 use Grav\Common\Uri;
 use Grav\Framework\Flex\Interfaces\FlexObjectInterface;
+use Grav\Framework\Session\SessionInterface;
 use Grav\Plugin\Form\Form;
 use Grav\Plugin\Login\Events\UserLoginEvent;
 use Grav\Plugin\Login\Login;
@@ -206,7 +207,7 @@ class LoginPlugin extends Plugin
 
             foreach ($pages->instances() as $page) {
                 $header = $page->header();
-                if (isset($header) && isset($header->access) && isset($header->login['visibility_requires_access']) && $header->login['visibility_requires_access'] === true) {
+                if ($header && isset($header->access) && isset($header->login['visibility_requires_access']) && $header->login['visibility_requires_access'] === true) {
                     $config = $this->mergeConfig($page);
                     $access = $this->login->isUserAuthorizedForPage($user, $page, $config);
                     if ($access === false) {
@@ -252,6 +253,8 @@ class LoginPlugin extends Plugin
                     $redirect = $page->route() . ($uri->params() ?: '');
                 }
             }
+        } else {
+            $redirect = $this->grav['session']->redirect_after_login;
         }
 
         $this->grav['session']->redirect_after_login = $redirect;
@@ -415,6 +418,7 @@ class LoginPlugin extends Plugin
                             $redirect_code = $loginEvent->getRedirectCode();
                         }
                     }
+                    $this->grav->fireEvent('onUserActivated', new Event(['user' => $user]));
                 }
             } else {
                 $message = $this->grav['language']->translate('PLUGIN_LOGIN.INVALID_REQUEST');
@@ -669,7 +673,6 @@ class LoginPlugin extends Plugin
         }
 
         $form->validate();
-        $form->filter();
 
         /** @var Data $form_data */
         $form_data = $form->getData();
@@ -825,7 +828,6 @@ class LoginPlugin extends Plugin
         $language = $this->grav['language'];
 
         $form->validate();
-        $form->filter();
 
         /** @var Data $form_data */
         $form_data = $form->getData();
@@ -1073,7 +1075,14 @@ class LoginPlugin extends Plugin
 
     public function userLogin(UserLoginEvent $event)
     {
+        /** @var SessionInterface $session */
         $session = $this->grav['session'];
+
+        // Prevent session fixation if supported.
+        // TODO: remove method_exists() test when requiring Grav v1.7
+        if (method_exists($session, 'regenerateId')) {
+            $session->regenerateId();
+        }
         $session->user = $event->getUser();
 
         if ($event->getOption('remember_me')) {
@@ -1102,6 +1111,10 @@ class LoginPlugin extends Plugin
             $login->rememberMe()->clearCookie();
         }
 
-        $this->grav['session']->invalidate()->start();
+        /** @var SessionInterface $session */
+        $session = $this->grav['session'];
+
+        // Clear all session data.
+        $session->invalidate()->start();
     }
 }
