@@ -23,7 +23,12 @@ $(document).ready(function(){
           newurl = newurl + "/hidepagetitle:true";
       }
 
-      window.location.href = newurl;
+      if (e.ctrlKey || e.metakey) {
+        window.open(newurl,'_blank');
+      } else {
+        window.location.href = newurl;
+      }
+
     }
   });
 });
